Allow syncing secrets into repository environment
This commit allows one to set an input called "environment", which when
enabled will sync the secrets as environment secrets to the same
repositories instead of putting the secrets as repository secrets.

The action will fail if the targeted repositories does not have the
specified environment created prior to running the action.