fix: change 'pageSize' inputs/args to 'per_page'
It's possible 'pageSize' works as intended, potentially due to some
kind of hard to find legacy support or overlap with octokit.net
(where the parameter name is explicitly that), but the correct name
of this parameter in the REST API is 'per_page'.

It appears likely that 'pageSize' has had no effect whatsoever. The
actual action code defaults the value to '30' which happens to align
with the REST API's default for when 'per_page' is not provided. While
the tests do use an explicit value (namely '3'), the Jest nock for each
endpoint dictate the number of results there instead of the actual
parameter value.